manage branches

+++ book-worm/src/main/java/org/example/dao/BranchDaoImpl.java
@@ -10,53 +10,55 @@ import java.util.List;
 
 public class BranchDaoImpl implements BranchDao{
     @Override
-    public boolean save(Branch branch) {
-        Session session = FactoryConfiguration.getInstance().getSession();
-        Transaction transaction = session.beginTransaction();
+    public boolean save(Branch branch, Session session) {
         Object save =  session.save(branch);
-        transaction.commit();
-        session.close();
         return save !=null;
     }
 
     @Override
-    public long count() {
-        Session session = FactoryConfiguration.getInstance().getSession();
+    public long count(Session session) {
         Query<Long> query = session.createQuery("SELECT COUNT(*) FROM Branch", Long.class);
         return query.uniqueResult();
     }
 
     @Override
-    public int getMaxUserId() {
-        Session session = FactoryConfiguration.getInstance().getSession();
+    public int getMaxUserId(Session session) {
         Query<Integer> query = session.createQuery("SELECT MAX(id) FROM Branch", Integer.class);
         Integer maxId = query.uniqueResult();
         return maxId != null ? maxId : 0; // Return 0 if no users exist
     }
 
     @Override
-    public void delete(int id) {
-        Session session = FactoryConfiguration.getInstance().getSession();
-        Transaction transaction = session.beginTransaction();
-        Branch branch = getDetail(id);
-        session.delete(branch);
-        transaction.commit();
-        session.close();
+    public boolean delete(String branchName, Session session) {
+        Query<Branch> query = session.createQuery("FROM Branch WHERE branchName = :branchName", Branch.class);
+        query.setParameter("branchName", branchName);
+        Branch branch = query.uniqueResult();
+        if (branch != null) {
+            session.delete(branch);
+            return true;
+        } else {
+            // Branch not found
+            return false;
+        }
     }
 
     @Override
-    public Branch getDetail(int id) {
-        Session session = FactoryConfiguration.getInstance().getSession();
-        Transaction transaction = session.beginTransaction();
-        Branch branch = session.get(Branch.class, id);
-        transaction.commit();
-        session.close();
-        return branch;
+    public boolean update(Branch branch, Session session) {
+        session.update(branch);
+        return true;
     }
 
     @Override
-    public List<Branch> getAll() {
-        return null;
+    public Branch getBranch(String branchName, Session session) {
+        Query<Branch> query = session.createQuery("FROM Branch WHERE branchName = :branchName", Branch.class);
+        query.setParameter("branchName", branchName);
+        return query.uniqueResult();
+    }
+
+    @Override
+    public List<Branch> getAll(Session session) {
+        Query<Branch> query = session.createQuery("FROM Branch ", Branch.class);
+        return query.list();
     }
 
 }
